Cleanup and document accessors package
