+++ geary-core/src/main/kotlin/com/mineinabyss/geary/ecs/accessors/types/ComponentOrDefaultAccessor.kt
@@ -1,18 +1,15 @@
-package com.mineinabyss.geary.ecs.accessors
+package com.mineinabyss.geary.ecs.accessors.types
 
+import com.mineinabyss.geary.ecs.accessors.ArchetypeCacheScope
+import com.mineinabyss.geary.ecs.accessors.RawAccessorDataScope
 import com.mineinabyss.geary.ecs.api.GearyComponent
 import com.mineinabyss.geary.ecs.api.GearyComponentId
 
-public open class ComponentAccessor<T : GearyComponent?>(
-    index: Int,
-    private val componentId: GearyComponentId,
-) : Accessor<T>(index) {
-    private val ArchetypeCacheScope.dataIndex by cached { archetype.indexOf(componentId) }
-
-    override fun RawAccessorDataScope.readData(): List<T> =
-        listOf(archetype.componentData[dataIndex][row] as T)
-}
-
+/**
+ * Implements the accessor [getOrDefault] operation.
+ *
+ * @see getOrDefault
+ */
 public class ComponentOrDefaultAccessor<T : GearyComponent?>(
     index: Int,
     componentId: GearyComponentId,
@@ -22,6 +19,7 @@ public class ComponentOrDefaultAccessor<T : GearyComponent?>(
 
     override fun RawAccessorDataScope.readData(): List<T> {
         if(dataIndex == -1) return listOf(default)
+        @Suppress("UNCHECKED_CAST") // Index assignment ensures this should always be true
         return listOf(archetype.componentData[dataIndex][row] as T)
     }
 }
