+++ geary-core/src/main/kotlin/com/mineinabyss/geary/ecs/entities/Relationship.kt
@@ -4,7 +4,6 @@ package com.mineinabyss.geary.ecs.entities
 
 import com.mineinabyss.geary.ecs.api.GearyEntityId
 import com.mineinabyss.geary.ecs.api.engine.componentId
-import com.mineinabyss.geary.ecs.api.engine.type
 import com.mineinabyss.geary.ecs.api.entities.GearyEntity
 import com.mineinabyss.geary.ecs.api.entities.toGeary
 import com.mineinabyss.geary.ecs.api.entities.with
