+++ geary-core/src/main/kotlin/com/mineinabyss/geary/ecs/helpers/ComponentHelpers.kt
@@ -1,7 +1,6 @@
 package com.mineinabyss.geary.ecs.helpers
 
 import com.mineinabyss.geary.ecs.api.GearyComponent
-import com.mineinabyss.geary.ecs.api.engine.type
 import com.mineinabyss.geary.ecs.api.entities.GearyEntity
 import com.mineinabyss.geary.ecs.api.entities.toGeary
 import com.mineinabyss.geary.ecs.prefab.PrefabKey
