+++ geary-core/src/main/kotlin/com/mineinabyss/geary/ecs/prefab/ParseRelationOnPrefab.kt
@@ -1,10 +1,10 @@
 package com.mineinabyss.geary.ecs.prefab
 
 import com.mineinabyss.geary.ecs.accessors.TargetScope
-import com.mineinabyss.geary.ecs.accessors.get
+import com.mineinabyss.geary.ecs.accessors.building.get
 import com.mineinabyss.geary.ecs.api.autoscan.AutoScan
+import com.mineinabyss.geary.ecs.api.autoscan.Handler
 import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.api.systems.Handler
 import com.mineinabyss.geary.ecs.components.RelationComponent
 import com.mineinabyss.geary.ecs.components.RelationOnPrefab
 import com.mineinabyss.geary.ecs.serialization.parseEntity
