+++ geary-core/src/test/java/com/mineinabyss/geary/ecs/accessors/AccessorHolderTest.kt
@@ -1,5 +1,8 @@
 package com.mineinabyss.geary.ecs.accessors
 
+import com.mineinabyss.geary.ecs.accessors.building.get
+import com.mineinabyss.geary.ecs.accessors.building.getOrDefault
+import com.mineinabyss.geary.ecs.accessors.building.map
 import com.mineinabyss.geary.ecs.api.engine.Engine
 import com.mineinabyss.geary.ecs.api.engine.entity
 import com.mineinabyss.geary.ecs.engine.GearyEngine
