+++ geary-core/src/test/java/com/mineinabyss/geary/ecs/api/entities/GearyEntityTests.kt
@@ -3,7 +3,6 @@ package com.mineinabyss.geary.ecs.api.entities
 import com.mineinabyss.geary.ecs.api.engine.Engine
 import com.mineinabyss.geary.ecs.api.engine.componentId
 import com.mineinabyss.geary.ecs.api.engine.entity
-import com.mineinabyss.geary.ecs.api.engine.type
 import com.mineinabyss.geary.ecs.components.PersistingComponent
 import com.mineinabyss.geary.ecs.engine.GearyEngine
 import com.mineinabyss.geary.ecs.engine.getArchetype
