+++ geary-core/src/test/java/com/mineinabyss/geary/ecs/query/GearyTypeFamilyExtensionsKtTest.kt
@@ -3,7 +3,6 @@ package com.mineinabyss.geary.ecs.query
 import com.mineinabyss.geary.ecs.api.engine.Engine
 import com.mineinabyss.geary.ecs.api.engine.componentId
 import com.mineinabyss.geary.ecs.api.engine.entity
-import com.mineinabyss.geary.ecs.api.engine.type
 import com.mineinabyss.geary.ecs.api.relations.Relation
 import com.mineinabyss.geary.ecs.api.relations.RelationValueId
 import com.mineinabyss.geary.ecs.engine.GearyEngine
